+++ RELEASES.md
@@ -2,6 +2,7 @@ Version 0.8.17 (2022-XX-XX)
 ========================
 * New driving model
   * Internal feature space accuracy increased tenfold during training, this makes the model dramatically more accurate.
+* Subaru Crosstrek Hybrid 2020 support thanks to martinl!
 
 Version 0.8.16 (2022-08-26)
 ========================